generate token for customer and admin
